requirement.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
langchain==0.2.16
langchain-groq==0.1.6
